Revert NamespaceChange change
