+++ test/kind.go
@@ -22,7 +22,6 @@ func init() {
 		"TestCorruptedTokenLogin",
 		"TestRunPrivateSource",
 		"TestEventsStream",
-		"TestPublicAPI",
 		"TestRPC",
 	}
 	maxTimeMultiplier = 3
@@ -125,95 +124,5 @@ func testDeleteOwnAccount(t *T) {
 	outp, err := cmd.Exec("auth", "delete", "account", "admin")
 	if err == nil {
 		t.Fatal(string(outp))
-	}
-}
-
-func TestPublicAPI(t *testing.T) {
-	TrySuite(t, testPublicAPI, retryCount)
-}
-
-func testPublicAPI(t *T) {
-	t.Parallel()
-	serv := NewServer(t, WithLogin())
-	defer serv.Close()
-	if err := serv.Run(); err != nil {
-		return
-	}
-
-	cmd := serv.Command()
-	err := ChangeNamespace(cmd, serv.Env(), "random-namespace")
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-	// login to admin account
-	if err := Login(serv, t, "admin", "micro"); err != nil {
-		t.Fatalf("Error logging in %s", err)
-		return
-	}
-
-	outp, err := cmd.Exec("user", "namespace")
-	if err != nil || strings.TrimSpace(string(outp)) != "random-namespace" {
-		t.Fatal(string(outp), err)
-		return
-	}
-
-	if err := Try("Find helloworld", t, func() ([]byte, error) {
-		outp, err = cmd.Exec("user")
-		if err != nil || strings.TrimSpace(string(outp)) != "admin" {
-			return outp, err
-		}
-		return outp, err
-	}, 10*time.Second); err != nil {
-		return
-	}
-
-	if err := Try("Find helloworld", t, func() ([]byte, error) {
-		return cmd.Exec("run", "helloworld")
-	}, 5*time.Second); err != nil {
-		return
-	}
-
-	if err := Try("Find helloworld", t, func() ([]byte, error) {
-		outp, err := cmd.Exec("status")
-		if err != nil {
-			return outp, err
-		}
-
-		// The started service should have the runtime name of "service/example",
-		// as the runtime name is the relative path inside a repo.
-		if !statusRunning("helloworld", "latest", outp) {
-			return outp, errors.New("Can't find example helloworld in runtime")
-		}
-		return outp, err
-	}, 15*time.Second); err != nil {
-		return
-	}
-
-	if err := Try("Call helloworld", t, func() ([]byte, error) {
-		outp, err := cmd.Exec("helloworld", "--name=joe")
-		if err != nil {
-			outp1, _ := cmd.Exec("logs", "helloworld")
-			return append(outp, outp1...), err
-		}
-		if !strings.Contains(string(outp), "Msg") {
-			return outp, err
-		}
-		return outp, err
-	}, 90*time.Second); err != nil {
-		return
-	}
-
-	if err := Try("curl helloworld", t, func() ([]byte, error) {
-		bod, rsp, err := curl(serv, "random-namespace", "helloworld?name=Jane")
-		if rsp == nil {
-			return []byte(bod), fmt.Errorf("helloworld should have response, err: %v", err)
-		}
-		if _, ok := rsp["Msg"].(string); !ok {
-			return []byte(bod), fmt.Errorf("Helloworld is not saying hello, response body: '%v'", bod)
-		}
-		return []byte(bod), nil
-	}, 90*time.Second); err != nil {
-		return
 	}
 }
